fraud_nlj_app/requirements.txt
flask
transformers
sentence-transformers
faiss-gpu
torch